[FEAT] 계정 관련 구조 변경
