+++ client/src/App.jsx
@@ -13,7 +13,6 @@ import {
   SignUp,
   Question,
   ProfileEdit,
-  RegisterProduct,
   Rank,
   GuideFaq,
   MyProfile,
@@ -24,6 +23,8 @@ import {
   NotFound,
   CategoryLayout,
   CategoryPosts,
+  ProfileLayout,
+  MyFavPosts,
 } from './pages';
 import { SIGNIN_PATH } from './constants/routes';
 
@@ -75,10 +76,7 @@ const router = createBrowserRouter([
         loader: postDetailLoader(queryClient),
         element: <Post />,
       },
-      {
-        path: 'myposts',
-        element: <AuthenticationGuard redirectTo={SIGNIN_PATH} element={<MyPosts />} />,
-      },
+
       { path: 'guide-faq', element: <GuideFaq /> },
       {
         path: 'question',
@@ -86,20 +84,30 @@ const router = createBrowserRouter([
       },
       { path: 'rank', loader: rankLoader(queryClient), element: <Rank /> },
       {
-        path: 'profile/:nickName',
+        path: 'profile',
+        element: <AuthenticationGuard redirectTo={SIGNIN_PATH} element={<ProfileLayout />} />,
+        children: [
+          {
+            index: true,
+            element: <MyProfile />,
+          },
+          {
+            path: 'edit',
+            element: <ProfileEdit />,
+          },
+          {
+            path: 'fav',
+            element: <MyFavPosts />,
+          },
+          {
+            path: 'myposts',
+            element: <MyPosts />,
+          },
+        ],
+      },
+      {
+        path: 'user-profile/:nickName',
         element: <Profile />,
-      },
-      {
-        path: '/myprofile',
-        element: <AuthenticationGuard redirectTo={SIGNIN_PATH} element={<MyProfile />} />,
-      },
-      {
-        path: '/profile/edit',
-        element: <AuthenticationGuard redirectTo={SIGNIN_PATH} element={<ProfileEdit />} />,
-      },
-      {
-        path: '/fav-category',
-        element: <AuthenticationGuard redirectTo={SIGNIN_PATH} element={<RegisterProduct />} />,
       },
       {
         path: '*',
