+++ client/src/components/profile/index.js
@@ -1,8 +1,7 @@
 export { default as MyProfile } from './MyProfile';
-export { default as MyProductList } from './MyProductList';
-export { default as Profile } from './Profile';
+// export { default as MyProductList } from './MyProductList';
 export { default as ProfileEditForm } from './ProfileEditForm';
-export { default as RegisterProductButton } from './RegisterProductButton';
+export { default as ProfileHeader } from './ProfileHeader';
 export { default as AvatarProfileInfo } from './AvatarProfileInfo';
 export { default as AvatarEditButton } from './AvatarEditButton';
 export { default as AvatarEditModal } from './AvatarEditModal';
