+++ client/src/pages/MyFavPosts.jsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const MyFavPosts = () => {
+  console.log('todo');
+  return <div>좋아요 글 목록</div>;
+};
+
+export default MyFavPosts;
