+++ client/src/pages/ProfileLayout.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Outlet } from 'react-router-dom';
+import { ProfileHeader } from '../components';
+
+const ProfileLayout = () => (
+  <>
+    <ProfileHeader />
+    <Outlet />
+  </>
+);
+
+export default ProfileLayout;
