+++ client/src/pages/index.js
@@ -7,10 +7,12 @@ export { default as MyPosts } from './MyPosts';
 export { default as Post } from './Post';
 export { default as CategoryPosts } from './CategoryPosts';
 export { default as GuideFaq } from './GuideFaq';
+export { default as ProfileLayout } from './ProfileLayout';
 export { default as Profile } from './Profile';
 export { default as MyProfile } from './MyProfile';
+export { default as MyFavPosts } from './MyFavPosts';
 export { default as ProfileEdit } from './ProfileEdit';
-export { default as RegisterProduct } from './RegisterProduct';
+// export { default as RegisterProduct } from './RegisterProduct';
 export { default as Question } from './Question';
 export { default as Rank } from './Rank';
 export { default as NotFound } from './NotFound';
